# sym_infer is noop for int input