no need to make the input a `NumNode` and get the int back, this is significant for jitted path because inputs to infer global_size, local_szie and ops are mostly int.

jitted llama on M1 Max `GPU=1 JIT=1 python -O examples/llama.py --prompt="Hello." --count=10 --temperature=0 --timing`

master
```
ran model in 6.49 ms
total 95.54 ms, 10.47 tok/sec
 year
ran model in 6.97 ms
total 95.47 ms, 10.47 tok/sec
 old
ran model in 6.47 ms
total 95.80 ms, 10.44 tok/sec
 male
```

this PR
```
ran model in 5.54 ms
total 94.95 ms, 10.53 tok/sec
 year
ran model in 5.45 ms
total 95.34 ms, 10.49 tok/sec
 old
ran model in 5.59 ms
total 95.64 ms, 10.46 tok/sec
 male
```